requirements.txt
Django==5.2.0
django-tinymce==3.6.1
django-import-export==4.3.1
djangorestframework==3.16.0
pillow==11.1.0
pytest-django==4.8.0
pytest-cov==4.1.0
factory-boy==3.3.0
pytest-factoryboy==2.5.1
drf-spectacular==0.28.0
waitress==2.1.2
psycopg2-binary==2.9.9
dj-database-url==2.1.0
whitenoise==6.6.0
graphene-django==3.2.1
